+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Details
   * Change vote results order [#234](https://github.com/demokratie-live/democracy-client/issues/234)
   * Fix positioning of vote results [#234](https://github.com/demokratie-live/democracy-client/issues/234)
+  * Change title for "in Vorbeeitung" [#252](https://github.com/demokratie-live/democracy-client/issues/252)
 
 ### 0.6.5